Implement button visibility toggling
This commit adds the ability to hide and show buttons through a runtime
command. This can be useful to futher immerse the player, or when the
author chooses to hide them out of a mechanical reason.

Note that this doesn't allow the hiding of the restart button for
multiple reasons. More testing must be done to make sure that the
restart button isn't used as often as I think.

Future commit should allow for a more seamless toggling. At the moment
we are wasting space with something that reeks of DRY.