Safeguard against invalid and/or mixed-case args
The user isn't expected to pass case-sensitive arguments in this
language; make sure we are processing these mixed-case situations, where
the user can just enter in similar (but not so) arguments.

PS1: Also make sure we handle potentially invalid args that can be
passed to changeBackgroundColor. We don't want any Scene vomit.