Unbreak delimiter splitting in scene_button
If a user passes the following arguments to scene_button:

"Hello, hello", hello

Then the function splitToArray will instead create an array with the
size of 3, causing the scene_button to fail. In addition, we won't have
"Hello, hello" as the button's text either, as it's cut by splitToArray.

To fix this, and to unify how functions with multiple arguments are
handled in general, change the common delimiter to a space and rework
how scene_button handles its arguments. From what I know, ChoiceScript
loves to abuse space anyway (see: rand, image, input, etc.), so this
isn't a grade A guess.

As a side effect of this commit, hide_button and show_button are also
affected.